NER merging functions are now method of DependencyTree.
