+++ ppp_questionparsing_grammatical/__init__.py
@@ -2,7 +2,7 @@
 
 from ppp_libmodule import HttpRequestHandler
 from .dependencyTree import Word, DependenciesTree, TreeGenerator, computeTree
-from .preprocessingMerge import mergeNamedEntityTagChildParent, mergeNamedEntityTagSisterBrother, mergeNamedEntityTag, QuotationHandler, preprocessingMerge
+from .preprocessingMerge import QuotationHandler, preprocessingMerge
 from .questionWordProcessing import identifyQuestionWord, questionWordDependencyTree, questionWordNormalForm
 from .dependencyAnalysis import simplify
 from .normalization import normalFormProduction
@@ -16,4 +16,4 @@ def app(environ, start_response):
     return HttpRequestHandler(environ, start_response, RequestHandler) \
             .dispatch()
 
-__all__ = ['DependenciesTree', 'computeTree', 'QuotationHandler', 'TreeGenerator', 'mergeNamedEntityTagChildParent', 'mergeNamedEntityTagSisterBrother', 'mergeNamedEntityTag', 'simplify', 'identifyQuestionWord', 'QuotationError', 'GrammaticalError', 'Nounificator', 'preprocessingMerge', 'normalFormProduction']
+__all__ = ['DependenciesTree', 'computeTree', 'QuotationHandler', 'TreeGenerator', 'simplify', 'identifyQuestionWord', 'QuotationError', 'GrammaticalError', 'Nounificator', 'preprocessingMerge', 'normalFormProduction']
